make it so `_parse_file` accepts multiple extensions
modify the `crowsetta.validation._parse_file` function
so that the type of the `extension` argument can be
not just a string but also a tuple of strings

this way other functions can specify a set of valid
extensions, e.g. `('.phn', '.PHN')`, not just `'.phn'`.

Makes it possible for e.g. `phn2annot` to act in a
"case-insensitive" way, converting both `.phn` and
`.PHN` files