TST/BUG: remove line from unit tests that caused error
A line in the unit tests for formats that tested
`format2csv` functions would pull of the "header",
but includes a row with actual data.

Specifically, we lose one row where we have an
`annot_path`.
If the first annotation file only has a single segment,
then that `annot_path` does not get added to the list
that we check to make sure all the files ended up in
the generated csv.

This didn't cause a test to fail before because usually
there were multiple rows in the first annotation file,
so the filename would still appear in the list of filenames
even when we threw the first segment away as a "header".

Removing the line in all the format tests
prevents this from happening.