ENH: Add functions to extract example data, fix #184
Adds functions that extract example annotation files
from the ``crowsetta.data.`` package
to the local file system.
This makes it easier to work with the
files in the way that ``crowsetta`` expects,
by passing in a path to an annotation file on the system.

The same thing cannot be achieved
by using ``importlib.resources``
because the abstractions in that package assume
we want to open files for the user,
and so we'll never need the actual files,
and can abstract them away in an API
that lets us work with them
even if they're in a compressed .zip, i.e. a .whl.

Without this change, ``crowsetta.data.get``
would have to return a context manager.
This could give users the wrong idea that they
**always** need to use a context manager,
i.e., if they learn to use the package
by reading examples in the docstrings.
By the same token,
it makes examples in the docstrings extra verbose.