Remove exceptions from birdsongrec.py
This removes NotADirectory and FileNotFound errors raised by
`birdsongrec2annot` when `wavpath` doesn't exist or it can't find
the wav files in the specified `wavpath`, respectively.

My intent was to make sure the files were there,
if something else depended on them.
But this prevents someone from accessing the annotations
if they don't have the wav files.
Which is the opposite of the point of this package.