Revert "Remove exceptions from birdsongrec.py"
This reverts commit c22804052e5509ca2ae027305544da15e18fee18.

We do actually need to check for audio with this
annotation format, since we use the audio
to get the sampling rate and convert the
onsets and offsets from sample numbers to seconds.
There might be a more flexible way to handle this long term,
but this did not fix the problem I had upstream with vak
so I'm just going to revert it for now.