Have SimpleSeq only map column names in `column_map`, fix #272
Inside `SimpleSeq.from_file`, when someone specifies a
`column_map` argument, we only do the mapping if a column
name in the loaded csv file is a key in `column_map`.
Otherwise we leave the column as-is, and if it's not one of
the columns we need for the annotations
("onset_s", "offset_s", "label"),
it gets thrown away.