Make SimpleSeq work with csv that doesn't have labels, fix #271
Adds `default_label` argument to `from_file` method,
a string that defaults to `'-'`
and if the `pandas.DataFrame`
loaded from the csv file does not already have a
'label' column, adds that column with the
`default_label` value as the label in every row.

This makes it possible to load segmentations saved
in csv files that do not have labels,
just start and stop times of the segments.